更新sdk，修复bug，增强协作工具

+++ MLN-Android/mlnservics/src/main/java/com/immomo/mls/fun/ud/UDPaint.java
@@ -119,7 +119,7 @@ public class UDPaint extends LuaUserdata<Paint> {
         if (values.length > 0) {
             LuaValue value = values[0];
             if (value.isNumber()) {
-                javaUserdata.setColor(value.toInt());
+                javaUserdata.setColor((int) value.toLong());
             } else {
                 UDColor color = (UDColor) values[0].toUserdata();
                 javaUserdata.setColor(color.getColor());
